backend/requirements.txt
django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
django-user-agents==0.4.0
django-ipware==5.0.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
python-decouple==3.8
Pillow==10.2.0
stripe==7.14.0
requests==2.31.0